SAAS-526 fix eslint errors
