+++ package.json
@@ -23,25 +23,26 @@
   ],
   "scripts": {
     "clean": "rimraf dist",
-    "start": "yarn clean && yarn dev",
+    "start": "npm run clean && npm run dev",
     "lint": "eslint --ext js,ts,tsx src",
-    "lint:fix": "yarn lint --fix",
+    "lint:fix": "npm run lint --fix",
     "dev": "tsc -p tsconfig.json --outDir dist --watch",
     "test": "react-scripts test --verbose",
     "test:ci": "CI=true react-scripts test",
     "test:coverage": "react-scripts test --coverage --watchAll=false",
-    "eject": "react-scripts eject",
-    "build": "yarn lint && yarn clean && tsc -p tsconfig.json --outDir dist",
+    "build": "NODE_ENV=production npm run lint && npm run clean && tsc -p tsconfig.json --outDir dist",
     "release": "np"
   },
   "dependencies": {
     "@popperjs/core": "^2.5.3",
     "react-popper": "^2.2.3",
-    "react-scripts": "3.4.3",
-    "typescript": "~3.7.5"
+    "react-scripts": "^4.0.3",
+    "typescript": "^4.1.3"
   },
   "devDependencies": {
-    "@grafana/eslint-config": "^2.0.3",
+    "@grafana/data": "7.1.3",
+    "@grafana/ui": "7.1.3",
+    "@grafana/eslint-config": "^2.3.0",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^10.4.9",
     "@testing-library/user-event": "^12.1.3",
@@ -51,16 +52,18 @@
     "@types/node": "^12.0.0",
     "@types/react": "^16.9.0",
     "@types/react-dom": "^16.9.0",
+    "emotion": "^10.0.27",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.2",
     "enzyme-to-json": "^3.4.4",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-config-prettier": "^6.12.0",
+    "eslint-config-airbnb": "^18.2.1",
+    "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-jest": "^24.1.0",
     "eslint-plugin-no-inline-styles": "^1.0.5",
-    "eslint-plugin-react-hooks": "^4.1.2",
+    "final-form": "^4.20.1",
     "jest-canvas-mock": "^2.2.0",
     "jest-enzyme": "^7.1.2",
+    "react-final-form": "^6.5.1",
     "rimraf": "^3.0.2"
   },
   "peerDependencies": {
