+++ src/components/Icon/types.ts
@@ -1,3 +1,5 @@
+import React from 'react';
+
 export type AvailableIcons = 'plusSquare' | 'minusSquare' | 'selectedSquare' | 'unselectedSquare';
 
 export type Icons = {
