+++ src/components/StepProgress/Step/index.ts
@@ -1,0 +1,1 @@
+export { Step, StepStatus } from './Step';
