+++ src/components/StepProgress/StepProgress.types.ts
@@ -1,0 +1,15 @@
+import { ReactNode } from 'react';
+import { FormRenderProps } from 'react-final-form';
+
+export interface StepProgressProps {
+  steps: OneStep[];
+  initialValues?: Record<string, any>;
+  onSubmit: (values: Record<string, any>) => void;
+}
+
+interface OneStep {
+  render: (props: FormRenderProps) => ReactNode;
+  title?: string;
+  fields: string[];
+  dataQa?: string;
+}
