+++ src/icons/ThemeDark.tsx
@@ -1,12 +1,10 @@
 import React, { FC } from 'react';
 
-export const ThemeDark: FC<React.SVGAttributes<SVGElement>> = (props) => {
-  return (
-    <svg width="26" height="26" viewBox="3 0 26 22" fill="none" role="img" aria-labelledby="title" {...props} xmlns="http://www.w3.org/2000/svg">
-      <title id="title">Dark theme</title>
-      <g stroke="currentColor" strokeLinecap="round" strokeLinejoin="round">
-        <path clipRule="evenodd" d="M13.3744 13.9592c-3.60274-3.6028-4.71502-8.68332-3.2485-12.90333-.34363.27153-.67668.56498-.99417.88247-4.49154 4.49154-4.49154 11.77196 0 16.26346 4.49157 4.4915 11.77187 4.4915 16.26347 0 .3175-.3175.6109-.6505.8832-.9935-4.2201 1.4666-9.3013.3536-12.904-3.2491v0z" />
-      </g>
-    </svg>
-  );
-};
+export const ThemeDark: FC<React.SVGAttributes<SVGElement>> = (props) => (
+  <svg width="26" height="26" viewBox="3 0 26 22" fill="none" role="img" aria-labelledby="title" {...props} xmlns="http://www.w3.org/2000/svg">
+    <title id="title">Dark theme</title>
+    <g stroke="currentColor" strokeLinecap="round" strokeLinejoin="round">
+      <path clipRule="evenodd" d="M13.3744 13.9592c-3.60274-3.6028-4.71502-8.68332-3.2485-12.90333-.34363.27153-.67668.56498-.99417.88247-4.49154 4.49154-4.49154 11.77196 0 16.26346 4.49157 4.4915 11.77187 4.4915 16.26347 0 .3175-.3175.6109-.6505.8832-.9935-4.2201 1.4666-9.3013.3536-12.904-3.2491v0z" />
+    </g>
+  </svg>
+);
